Add Levenberg damping to LM strategy.
Previously, Ceres only supported the Marquardt damping, which is widely
known as diag(J'*J) where J is the Jacobian. This commit allows user to
choose between the Levenberg damping, which is just an identity matrix,
and the Marquardt damping. (Sometimes, the Levenberg damping shows
better rank-filling behaviour than the Marquardt one.)  For
compatibility, Marquardt damping is used by default.

Change-Id: Ideaf90f418310e4d20a4e3a9e8c90b6c3d532081